wip - Rename 'foundation' package to 'system' to avoid OMF "Foundation" overload
