+++ packages/overture-schema-addresses-theme/src/overture/schema/addresses/address/models.py
@@ -9,12 +9,12 @@ from overture.schema.core import (
     StrictBaseModel,
 )
 from overture.schema.core.types import CountryCode
-from overture.schema.foundation.primitive import (
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
 )
-from overture.schema.foundation.string import StrippedString
+from overture.schema.system.string import StrippedString
 
 
 class AddressLevel(StrictBaseModel):
