+++ packages/overture-schema-base-theme/src/overture/schema/base/land/models.py
@@ -11,7 +11,7 @@ from overture.schema.core import (
     Feature,
 )
 from overture.schema.core.models import Named, Stacked
-from overture.schema.foundation.primitive import (
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
