+++ packages/overture-schema-base-theme/src/overture/schema/base/land_cover/models.py
@@ -9,7 +9,7 @@ from overture.schema.core import (
     Feature,
 )
 from overture.schema.core.models import CartographicallyHinted
-from overture.schema.foundation.primitive import (
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
