+++ packages/overture-schema-base-theme/src/overture/schema/base/models.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 from overture.schema.base.types import SourceTags
-from overture.schema.foundation.string import WikidataId
+from overture.schema.system.string import WikidataId
 
 
 class SourcedFromOpenStreetMap(BaseModel):
