+++ packages/overture-schema-buildings-theme/src/overture/schema/buildings/building/models.py
@@ -6,7 +6,7 @@ from pydantic import ConfigDict, Field
 
 from overture.schema.core import Feature
 from overture.schema.core.models import Named, Stacked
-from overture.schema.foundation.primitive import (
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
