+++ packages/overture-schema-core/src/overture/schema/core/validation.py
@@ -1,4 +1,4 @@
-from overture.schema.foundation.constraint import UniqueItemsConstraint
+from overture.schema.system.constraint import UniqueItemsConstraint
 from overture.schema.validation import (
     ConstraintValidatedModel,
     any_of,
