+++ packages/overture-schema-system/src/overture/schema/system/primitive/bbox.py
