+++ packages/overture-schema-system/src/overture/schema/system/string.py
@@ -2,7 +2,7 @@ from typing import Annotated, NewType
 
 from pydantic import Field
 
-from overture.schema.foundation.constraint import (
+from overture.schema.system.constraint import (
     CountryCodeConstraint,
     HexColorConstraint,
     JsonPointerConstraint,
