+++ packages/overture-schema-system/tests/primitive/test_geometry.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel, ValidationError
 from pytest_subtests import SubTests
 from shapely import wkt
 
-from overture.schema.foundation.primitive import (
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
