+++ packages/overture-schema-transportation-theme/src/overture/schema/transportation/segment/models.py
@@ -10,8 +10,8 @@ from overture.schema.core import (
 from overture.schema.core.models import (
     Named,
 )
-from overture.schema.foundation.constraint import UniqueItemsConstraint
-from overture.schema.foundation.primitive import (
+from overture.schema.system.constraint import UniqueItemsConstraint
+from overture.schema.system.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
