+++ packages/overture-schema-validation/src/overture/schema/validation/types.py
@@ -4,7 +4,7 @@ from typing import Annotated, NewType
 
 from pydantic import Field
 
-from overture.schema.foundation.primitive import float64
+from overture.schema.system.primitive import float64
 
 from .constraints import (
     CategoryPatternConstraint,
